+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # CHANGELOG
 
 ### HEAD
+* handle changes to devicePixelRatio
+* improve debugging information in *.dev.js (Test wether calculated sizes matches rendered size of image + Test order of source elements and sizes attribute)
 * Stick to editorconfig in all files
+* removed UMD (simply not so good for polyfills, sorry)
 * Add `.editorconfig` file to ensure coding conventions
 * Update npm dependencies
+* add ``lazyload`` check
 
 ### v0.9.3-RC1
 ### v0.9.3